requirements.txt
youtube-transcript-api
langchain
langchain-community
langchain-google-genai
faiss-cpu
tiktoken
python-dotenv
streamlit